Finished!
Some things to note:
Steel was both reporting the wrong values for the breakdown. Now they're fine.
Knowledge doesn't apply Tax, Hunger, or the Global Mult to the Sundial, which philosophically makes sense but complicates the breakdown when attempting to calculate by hand. I moved the Sundial to after those line items in the breakdown to simplify it, but the Plasmid mult will confuse it again. Probably not a big deal; just something to be aware of.