Kinda bug fixes
If the Market tab is not unlocked yet and Storage is, Resources defaults to Storage instead of showing a blank Market tab.
Adding Unified trait enacts permanent Unification without any achievement unlocks.
Verrazzano perk description says the correct amount of extra planet choices that it provides and added the semi-colon.