Set of Bug Fixes
Launch Facility: The Launch Facility cannot be queued twice and will be removed from the queue if finished manually.

Perks: Listing of perks and the effects of many perks now check for a base achievement level of 1, to protect against the case where one is earned only in a Micro universe.