Autodeoccupation Addition/Fix + Build
Currently, when occupying Foreign Powers and then turning off buildings that provide Soldiers, the Garrison can end up with a negative amount of Soldiers. This causes numbers to jump back and forth between negative and 0, and occasionally causing the Wounded to irreversibly turn negative.
The autodeoccupation system will automatically deoccupy a Foreign Power(s) if the Soldiers they are using would make the number of Soldiers in the Garrison turn negative.