Wiki Achievements Universe Filtering + Build
achievePage() function now filters achievements based on universe using an optional parameter for the desired universe.
(idk how to get the HTML to execute the idea, but the function itself works just fine)
Refactored format_emblem function to make use of a helper method to reduce repetitive code and work with universe filtering.