String Packs
Implemented the ability to import and use "String Packs". They're basically like resource packs in Minecraft. Just a way for people to customize the look of their game more, purely visual. It essentially allows for a locally sourced language file that sits on top of the language chosen, just as non-English languages sit on top of English, which serves as the basis. Wiki entry (Mechanics -> String Pack) explains in detail what one is and how one can be made.
As a plus, it could be a good vehicle for promoting the game and bolstering the community, as I'm sure themed packs will be worked on and shared, whether on reddit or discord.