Custom Lab Upgrades
Added a "Clear Fields" button that empties all text fields, resets the genus to the default, and unselects all traits to make creating new customs easier.
Added "Import Species" and "Export Species" buttons that allow for importing/exporting customs from/to .txt files to allow for easier sharing of customs and an easier way for people to save a custom for later use.