Hell Training Timer + Minor Fixes
Priests now appear on purchasing a Temple/Ziggurat, not seeing the tooltip.
Spy Operator config now uses the pointer for the cursor instead to indicate it can be dragged.
The Message Log can't be resized if it's on the right now, since it just gets auto-sized back to fill the page.
Flipped Micro Dark Energy calculators so homeworld/space cost creep use the right calculations now.
Removed useless code.
Added the Soldier training timer to the Hell Dimension tab.