Custom Lab Stuff + Build
Added page for the Custom Lab in the Wiki -> Species tab so people can fiddle around with customs without having to import a save before Ascension, and so people can experiment with different perk levels of Technophobe and/or Ascended.

Fixed bug where a person could import a custom that used traits that they didn't have unlocked and still be able to use the custom with those traits. Now it will remove any traits not yet unlocked.
Added Untapped Potential display next to number of free Genes in the custom lab, so people can see how much of a bonus they'll get.