Queue Behavior Changes/Fixes + Build
In ordered queues, Green Nevers (which happen when you have the potential cap for a resource, but don't have enough of it currently and don't have the production to ever produce it) now will not be barriers in the queue, freezing the queue on them and preventing the things after it from being completed. Now, they will act just like Red Nevers (which happen when you don't have the potential cap for a resource), and the queue will skip past them and move to the next item.

Fixed bug where in unordered queues an ARPA project at the end of the queue with a Never completion would block the rest of the queue from being completed. Fixes #488 Fixes #613

Fixed bug where the time estimations on buildings/techs could calculate an actual number as an estimate instead of Never when the building would in fact never complete. This was caused by the timeCheck functions proceeding to check the times for all the resources that something costs, even after it determines that it will never complete from one resource. So if any resource after that one that never completes does complete and was the determined bottleneck for that building/tech, then the -9999999 time was overridden by that bottleneck's time. Now if it determines that something will never complete, it skips past the remaining resources.

Added Play/Pause buttons next to the building queue and research queue titles that let you pause and resume each the queues.

Updated Queue Mechanics entry with info about queue merging options and pause/play.