Add "Dread Not" feat
The achievement "Dreaded" is earned if you ascend without owning
a dreadnought. It's possible to earn this achievement if you've
constructed a dreadnought but it was destroyed in an assault. The
description of the achievement doesn't make this obvious.

Introduce a new feat, "Dread Not" which is earned only if you ascend
without building even a single Dreadnought. Update the description of
Dreaded to mention that you must ascend without owning any Dreadnoughts.