refactor of accelerated time calculation
The `global.settings.at` variable was turned into units of milliseconds
and the hardcoded factor of 2.5 all over the place was removed.
`global.settings.gameSpeed` and `global.settings.atMultiplier` were
introduced and the timers for the game loops were added to `webWorker`
to calculate the correct factor where needed.

`atrack` variable was removed since it was equivalent to
`global.settings.at` and therefore redundant.

Gamespeed was added to topbar and removed from tooltips.

The `atMultiplier-1` is necessary in the topbar display as well as in
`longLoop()` in `src/main.js` because while the game runs at n-times
speed, only (n-1)-times are actually the bonus speed and the remaining
1-times is the normal game which should not count against the bonus
time.