Add biome restrictions to cold snap and heat wave
Certain biomes enforce temperature characteristics, so this seems like an oversight.

Volcanic and ashland biomes are "always hot" in the summer, so prevent cold snap at that time. Tundra and taiga biomes are the opposite, "always cold" in the winter, so prevent heat wave at that time.

Hellscape is "never cold," so prevent the cold snap event with no regard to the current season. Allow the event on a Permafrost Hellscape (custom) planet as an exception.

Eden is idyllic, so it normally can be hot only in summer (typically non-winter) and cold only in winter (typically non-summer). To preserve a one-step improvement in seasonal temperature range versus standard temperature biomes, prohibit cold snap in summer and heat wave in winter.

Continue to allow winter heat waves and summer cold snaps for normal biomes.