Convert slaves into a normal resource
Currently, slaves are stored as a count on the slave pen structure, and they are mirrored as a global resource for display purposes. This causes a value mismatch when a tax riot occurs, where slaves appear to be killed but are not. Change all references to the slave pen version of slave count into the global resource version.