Alphabetize bad genes
It is easier to tell if a bad gene is in the list or not when the list is alphabetical in the default language, English. The sequence will appear to be randomly shuffled for other languages, but that should already be the case now.

In addition, remove the second instance of invertebrate from the list, thereby reducing its probability to the same chance as other negative traits.