Fix priority inversion of DNA sequencer and build queues
The conditions for the DNA sequencer and the build/research queues to consume the Knowledge resource are different. As a result, structs and techs that require Knowledge may never build when automatic sequencing is enabled. Specifically, this may occur when Knowledge production less than 10k per day.

To allow queues to function for structs and techs near the Knowledge cap at that time, set the DNA sequencer buffer size to the lesser of 10,000 and the Knowledge production value.

Fixes issues #322 and #672.