Add setting to auto-power all new structs (#1010)
* Fix return status of powerOnNewStruct

Move "return true" inside of the check for power. Also, update the current available power status even when matter replicator governor task is off, so that future invocations are also correct.

* Always set powered() to electricity cost

For all structures that use the powered() function as a proxy for enabling ON/OFF buttons, return 0 instead of a misleading positive value

* Add button to always power new structs

Extend the functionality of powerOnNewStruct to handle both electricity and region support. Return true iff the new building is powered on.

To allow region support structure detection, add new field "s_type" to structures that consume support and have ON/OFF buttons. At this time, the field is not added to structures without ON/OFF buttons, such as Swarm Satellite.

Resolve incongruities in existing structure definitions:
- Add support() function to the zoo
- Check for available support before enabling Tau Ceti mining ships and whaling ships