Prevent exploits to assign too many crafters
Scarletite and Quantium limit the number of craftsmen working on them
based on the number of associated buildings that are currently
operating. Scarletite crafters are intended to be removed when an
Infernal Forge is powered off, but if the game is paused, then they can
be reassigned before the crafting window is reloaded. Also, crafters may
be assigned even when the forge is not
receiving enough power. Quantium
crafters are never forcibly removed
when a lab is powered down.

This change introduces a check for job limits on specific types of
craftsmen. It directly precedes the overall job limit, which can remove
only craftsmen that exceed the overall craftsman limit.