Support automatic power-on option for additional structures
The following structures are expected to automatically power on when
built, but the automatic power-on switch toggle (#1010) does not cause
them to ignore power requirements.

False path:
- Stargate

True path:
- AI Supercore
- Tau Ceti: Orbital Station (when dismantling the explorer ship)
- Tau Ceti: Ore Refinery

I tested all 4 changes individually.

Other structures that always require manual intervention, such as the
World Collider and the Ascension Machine, are not affected.