Screen reader: make more menus usable (#1111)
The Truepath ship constructor, mech constructor and industry tab > matter replicater menus all have to be hovered on with the mouse in order to open them. Screen readers have a key to make that happen, but it must often be pressed multiple times and the mouse can somehow move away while you're scrolling through the menu, closing it and making you try again.  This allows pressing enter on such a button to open its menu. Hovering still appears to work.

Fixes #1065