Improve tab refresh behavior (#1125)
* Improve tab refresh behavior

The code to refresh governor actions generally assumes that it is being executed on the tab where the action is made visible. However, Evolve has building and research queues, so these triggers might happen while the GUI is viewing any tab. Update the restrictions so that governor tasks are immediately refreshed even without Preload Tab Content enabled.

The defineGovernor() function now accordingly checks for its own visibility before it redraws the GUI.

Some completely missing governor tab refreshes, such as when adding or removing Slavery or Cannibalism, or when researching certain technologies, have also been added.

Additionally, remove various instances where the resource panel was deleted and redrawn, which seems to be legacy that isn't required anymore. The simple action of unlocking crates / containers is good enough to make those resources visible, to add the + buttons for per-resource storage management, and for the associated storage type to appear within the modal.

Containers may be unlocked by any building that grants them, because, unlike crates, containers aren't mandatory to progress in the game. Add a variety of checks for this, so even a user whose first container is provided by a building other than the container port or wharf can get the proper unlocks.

* Fix tab condition check when learning Personal Assistant