Fix scientific notation; use locale number formats (#1208)
* Fix scientific notation; use locale number formats

Evolve's default number display mode, SI, is actually engineering notation, because it uses only exponents that are multiples of 3. In addition, it can't handle exponents larger than 27.

Fix these deficiencies:
- Change scientific notation to use any integer exponent.
- Create a new engineering notation selection for users who like the current behavior.

For quite large numbers (1e27 and larger), metric suffixes become inadequate.
- Numbers 1e27 or larger use SI instead of a suffix.

Locale-aware number formatting is available in Hell observation statistics, but absent in the rest of the game. This is a good opportunity to provide locale-aware number formatting to more places.
- Use locale-aware formatting in sizeApproximation()
- Change "fixed" (always 1 fractional digit) to "precise", which enforces a minimum precision and displays at least all integer digits. This mode is used for Hell observations.
- Add new "exact" parameter that forces all fractional digits to display, even if the number gets really big. This is used for prestige resource counts.

* Allow +1 sigfig for single leading digit

Displayed number of digits is the same for leading zero and leading nonzero, even though apparent significant figure count is different

* More accurate comment

* Fix force SI logic to really be SI (not Eng)