Scale surveyors and gun emplacements for high population (#1102)
When the high population trait is in effect, gun emplacements now reduce the number of soldiers required to guard the soul forge by a scaled value. This addresses #1094.

This change also adds high population scaling to Hell Surveyors in the following ways:
- Job count and maximum carport damage scaling based on high population value (usually 4x)
- Carport repair scaling based on growth value (usually 3.5x)
- Surveyor danger scaling based on high population value (usually 4x)
- Infernite output is worker scaled (usually 0.26x) for high population
- Maximum demon bodies searched for soul gems by each surveyor are worker scaled down from 100 bodies (usually to 26 bodies)

Separately, the method of distributing predator drone kills to surveyors is slightly modified for improved accuracy.

This change will vastly increase surveyor deaths for insectoid races, so it's recommended to merge this together with or after PR #1082. I tested all of the changes, and each change appears to work as expected.

Co-authored-by: Peter Motschmann <pmotsch3@gmail.com>