Geothermal smelter related fixes (#1271)
Automatically smelt iron in geothermal plants when no homeworld is available. Prevent geothermal plants from smelting when they are not receiving enough He-3 fuel.

Add a smelting() method to all structures that provide smelting. Use it for all references to number of supported smelters.

Create common functions to enable a new smelter and to check whether smelting is possible.