Fix aria-labels for the mass ejector and matter replicator..
The "eject more/less" and "transmute more/less" buttons used the loc() function to get the localized names of the resources, but this didn't work properly with any resources with changed names. Here I just changed the calls to loc() to instead get the current names from the global.resource object, which seems to contain the localized names anyway.