Fix ARIA Labels Referencing Resources (#1360)
* Fixed several button aria-labels to use the correct resource name.

Changed the labels for the following buttons: the import and export buttons on the market tab, the add/remove crate/container buttons on the storage tab, the "Open crate management for ..." buttons next to each resource, and the add/remove crafter buttons on the civics tab.

* Change the resRate method to use the correct resource name for aria-labels.

* Changes aria-labels for the add/remove crafter buttons.

"Craftsman" were renamed to "crafters" in v1.3.17, but the aria-labels for assigning/unassigning them under the civics tab still said "Craftsman"

* Fix aria-labels for the mass ejector and matter replicator..

The "eject more/less" and "transmute more/less" buttons used the loc() function to get the localized names of the resources, but this didn't work properly with any resources with changed names. Here I just changed the calls to loc() to instead get the current names from the global.resource object, which seems to contain the localized names anyway.