Improve Accessibility of Message Queue Filters (#1376)
* make screen readers think filters are buttons

Added the attribute `role="button"` to the filter selections generated in index.js, this will force screen readers to treat all of them as separate clickable elements.

* add aria to distinguish selected filter

Now screen readers will detect if a filter is selected or non-clickable. Added the `aria-disabled` attribute to the filter elements. Edited the `swapFilters()` method to toggle the value between true and false when the `is-active` class is added or removed.