Racconar did not start with forager. (#1390)
* Fix for combining traits.

Renamed inactive to inactiveTraits as inactive is already used for isolation.

Fixed inactive value being added to the save on a game version that's too old.

* Fix racconar not starting with forager

Old way of giving Racconar forager was left in, which clashes with the new way of granting forager, causing it to remove forager instead.

* Fixed double initialization of inactiveTraits

* Moved inactiveTrait initialization back to where it was before