Add more structs to alwaysPower setting (#1384)
* Add more structs to alwaysPower setting

Centralize power_reqs checks to `powerOnNewStruct()`.

Allow mills to be automatically powered on after researching Wind Turbine.

Automatically power on mines, coal mines, rock quarries, cement plants, and metal refineries when the alwaysPower flag is set. Wardenclyffes and sawmills already did this, so no change to those. In all cases, the setting takes effect when the building is constructed, not when the technology is researched.

Only when the alwaysPower flag is enabled, automatically power on the completed World Collider, Stargate, Ascension Machine, Devilish Dish, or Waygate.

Force the Guard Post and Cultural Center to respect the usual rules for powering on. If the alwaysPower switch is disabled and power is insufficient, then no longer automatically power them on at time of completion.

Add initStruct-compatible definitions to some more structs.

* Actually check alwaysPower for ascension machine