Improve Accessibility of Manual Crafting Buttons (#1397)
* add button roles to manual craft buttons

Adds role="button" to the elements representing buttons to craft certain amounts of a resource or to craft as much as possible. Now, screen readers correctly interpret each as a single clickable button element.

* fix label of max craft buttons to use correct resource name

The "max craft" buttons still used "${name}", which results in an incorrect label if the resource in question is named something else currently. Replaced the substitution with "global.resource[name].name" to use the correct current name. (the labels for the other crafting buttons were previously fixed)