Miscellaneous fixes (#1399)
* Miscellaneous fixes

Do not nullify global production bonuses to uranium ash.

Do not provide quadratic global production bonuses to teamster income,
which is already derived as a percentage of other income sources that
include a global production factor.

Display base production speed in the Mining Outpost tooltip.

Fix some errors in strings.

Limit the maximum penalty from heat and cold intolerance to 100%.

Begin the anarchy penalty from the 11th citizen, matching the
description that is provided on the wiki. Scale high population such
that the number of penalty-free citizens matches the normal population
case, rather than reducing the value by 1.

Fix max structure count of Fire Support Base on the wiki.

Allow batteries in heavy mechs to work on the Demon Lord.

Clearly state in the wiki that multiple copies of a mech's special gear
item will not stack with each other.

Fix some cases where the wiki failed to render the full page

* More effective fix for morale tooltip

* Fix rounding precision error on Bunker