fix out of date reset list in wiki
Previously, the top of the wiki page covering resets stated that there were currently 9 reset options and listed 9 reset types. This information appears to be out of date; specifically, the count and list were missing the new T6 (Apotheosis) reset, as well as the 3 TP4 resets (Matrix, Retirement, and Garden of Eden). This change adds the 4 missing resets to the array defining this list. I have also rearranged the order of the array to match the sequence in which resets are covered: "vacuum" was moved after "blackhole", "terraform" was moved to before "infusion", and all of the True Path-specific resets were grouped at the end.