fix smelter aria-labels when evil
Usually, basic smelters by default burn lumber for fuel, but if the player's race is evil, the resource used will be changed to either flesh or souls (if playing a non-Wendigo soul eater race). Previously, the aria-labels associated with this resource always referenced "lumber fuel" or "wood fuel". This change fixes the labels for the "Add [Resource]" and "Remove [Resource]" buttons and adds a special case to the smelter's `ariaCount()` method to determine the correct resource name to be used when giving a count of smelters using that resource.