Force spy actions to always take time
Currently, spy actions fail to occur when the timer reaches 0 (happens
at 6 spies for influence ops with Spy Gadgets researched) and completely
break that nation's spy actions if the timer becomes negative.

As a mitigation, cap the number of spies that can reduce the duration of
a spy action to 1 fewer spy than would cut the timer to 0 time.

It's currently not possible for global spy action speedup (governor
efficiency upgrade, Dryad's befuddle trait, and Dryad thralls) to reduce
the minimum-valued 50 second timer to 0 time or less.