use correct project name for queue labels
The buttons to add an ARPA project to the building queue used the internal identifiers instead of the projects name. This would mean the label would say something like "Queue lhc" instead of "Queue Supercollider". This changes the label to use the "title" variable instead of"project", which contains the correct and localized name.""