combine project descriptions into one button
Previously, there were two separate clickable elements for each ARPA project to separately speak the project's description and effect strings: specifically, the element with the name of the project and the one showing the current level of the project (both screen reader only). I thought this was confusing, so I combined the "srDesc" and "srEffect" methods into one to speak both, and I changed the first element to use the combined method, added role="button", and included "description" in the label so it appears similar to the other description buttons in the game. The element showing the project level was changed to be non-interactive when clicked and was converted from an anchor to a span element so screen readers do not announce it as a link or "clickable".