Better handling of partial trade route fulfillment (#1428)
For trade routes that export a resource: fix a bug where the money for
all allocated export routes will be given when the available resource
supply can satisfy just 1 route. Only routes that can be filled will be
exported, and only those routes will give money.

For trade routes that import a resource: when stored money is enough for
at least 1 route but not enough for all allocated routes, import the
amount of routes that can be paid for, rather than nothing at all.