ARPA accessibility improvements (#1432)
* use correct project name for queue labels

The buttons to add an ARPA project to the building queue used the internal identifiers instead of the projects name. This would mean the label would say something like "Queue lhc" instead of "Queue Supercollider". This changes the label to use the "title" variable instead of"project", which contains the correct and localized name.""

* combine project descriptions into one button

Previously, there were two separate clickable elements for each ARPA project to separately speak the project's description and effect strings: specifically, the element with the name of the project and the one showing the current level of the project (both screen reader only). I thought this was confusing, so I combined the "srDesc" and "srEffect" methods into one to speak both, and I changed the first element to use the combined method, added role="button", and included "description" in the label so it appears similar to the other description buttons in the game. The element showing the project level was changed to be non-interactive when clicked and was converted from an anchor to a span element so screen readers do not announce it as a link or "clickable".

* add ARIA headings for each project

The Projects tab lacked headings for each project, which made it more annoying to navigate with a screen reader. This update modifies the spans containing each project title (which were previously hidden from screen readers), removes aria-hidden="true" so they can be detected, and assigns role="heading" with aria-level="3" to make them appear as level 3 headings to screen readers.

* add headings to Genetics tab

The Genetics tab also lacked navigational headings. This change adds ARIA roles to make the section labels for species' minor traits, major traits, and available traits (when unlocked) into heading level 3s. Screen-reader-only h4 elements were also added before each minor trait using the "is-sr-only" class. The traits in the other two sections are not navigated between often enough to require subheadings.