show cost and cost creep headings as h3 in wiki
All the "Cost" and "Cost Creep" headings for each structure and ARPA project were H2 headings, the same as the heading with each structure's name, so it could take up to three times longer to navigate one of these wiki pages using a screen reader. This change uses the aria-level attribute to make these headings appear as H3 headings to screen readers.