fix bug in removeTask function
The for loop in removeTask() tried to directly get the length of "global.race.governor.tasks", but as "global.race.governor.tasks" is an object and not an array, this didn't work and prevented the function from actually clearing disabled tasks. This change wraps the object in a call to "Object.keys()" to properly count the number of keys.