Fix Clearing of Disabled Governor Tasks (#1471)
* fix bug in removeTask function

The for loop in removeTask() tried to directly get the length of "global.race.governor.tasks", but as "global.race.governor.tasks" is an object and not an array, this didn't work and prevented the function from actually clearing disabled tasks. This change wraps the object in a call to "Object.keys()" to properly count the number of keys.

* add several redraws of governor tab after removing tasks

* remove ADV. Spy Operator with other spy tasks

The three main cases where spy tasks are removed (after unifying, adding the Unified trait, or when the jump gate shutdowns in True Path) only the standard spy recruiter and spy operator tasks were being removed. This adds lines to remove the ADV. spy operator task as well.]