fix static aria-labels for ARPA project buttons (#1467)
The fix in #1432 that replaced the internal project name with its actual project name malfunctions when a new monument is built and changes type; the labels still use the wrong old name as they were written as static labels. This change updates the "aria-label" for the project's queue and description buttons to make them reactive and update when the name changes. I had to add a new "projectName()" method along with the "loc" function to the Vue instance to get it working.