Redraw city and research tabs on governor change
Currently, when a governor is either appointed or fired, the code checks the building queue to rename any amphitheater/stadium or apartment/mansion items to update their titles if applicable. However, it does not update the name in the city and research tabs (for the apartment/mansion tech). This change adds redraws to the city and research tab and tech list when a governor is fired or appointed, and updates the first argument in the existing "updateQueueNames()" calls to true so that they also update the research queue along with the building queue.