Fix fortress and garrison troop requirements in SR desc
The code that lists requirements for structures or researches for their screen reader descriptions did not handle the "hellArmy" requirement (soldiers in the fortress), first used in the "Secure the Pit" structure and later used for a few more progression structures in the Elysium Fields. This broke the descriptions for these structures and caused an error when the game attempted to speak them.

2026-08-07: Corrected the fix to the non fortress troop requirement, before the lines to declare the type's label and the one to add the "insufficient resource" note were swapped.