log servant gains plus refactoring
As the number of gained servants and skilled servants are not included in the data from calcPrestige, separate calculations had to be added using code based off of the block updating the counts of servants at the start of a game in main.js. New strings for singular "Servant" and "Skilled Servant" were also added.

Also refactored code to figure out the correct name for the current resource in the report to be a switch statement.