add new setting to control prestige logging
Adds a new control to the settings tab to control whether prestige summaries should be logged. The new prestige messages will only be sent if the option is enabled (by default it's off). Put it by "disable special events" as that was the only option that seemed vaguely related to messages.

2026-07-07: Fix merge conflicts with latest working branch, in particular changed new setting to be "settings19".
2026-08-22: Bump setting to 20 after new solar names setting.