add new setting to control prestige logging
Adds a new control to the settings tab to control weather prestige summaries should be logged. The new prestige messages will only be sent if the option is enabled (by default its off). Put it by "disable special events" as that was the only option that seemed vaguely related to messages.