try making prestige messages easier to translate
Rephrased strings to have the reset type be at the start of the sentence. For example: `Ascension earned you 2225 Plasmids, 139 Phage, and 2.4 Harmony Crystals.` I just left `prestige_report_list_end` to hold a '.'.

Separated the string for " and" into two, one with the comma and one without. It is hoped that this will make it easier for languages which never have a comma (or other separator) before the and.

Also made some spacing at the top of code blocks slightly more consistant.