make city h2 reactive
Edit the code for the new city tab heading again so that the name correctly updates with the growing population reactively. The last commit made it updatable, but it would only do so when the tab reloaded.

Code wise, just added a call to vBind to the heading element and the tabLabel function.