Avoid overflow in foreign power level outputs
When the text reads 'Above Average (137)', it does not fit within 8rem and
therefore wraps to two lines. There is enough space for this text to be wider
within the layout. (Tested at various widths and layout break-points.)

8rem is too narrow.  9rem is too narrow.  10rem is just wide enough, but does
not have much room to grow.  12rem was chosen as a value which fits well in the
layouts available (at different break-points), and might work for other
languages too.  I have not tested with languages other than English.