add buttons to speak planet descriptions
Adds a visually hidden button on the planet selection screen that speaks the full planet description with the same information that appears in the visual tooltip including biome, orbit, traits, and geology when activated. Also added a line of code to refresh a button when its planet is GECKed.

A new function `srPlanetDesc` was added to actions.js to generate the spoken description with a few formatting changes from the original `planetDesc` so it speaks better (in particular, geologic deposits are spoken in a comma-separated list).