improve aria-labels for ship movement buttons
Add the target system's name to the ARIA labels for the buttons used to move ships within Andromeda. Which is to say, the labels now indicate the system a ship will be removed from or added to. For example, "Add Scout Ship to Gorddon System" or "Remove Dreadnought from Chthonian". The labels also now use the proper names of the ships instead of their internal identifiers.